+++ backend/src/services/user-service.js
@@ -159,7 +159,7 @@ function decrypt(text) {
  * @param {string=} additionalFields Fields to request from database separated by a single space
  * @returns {Promise<User?|Error>}
  */
-async function getByID(id, additionalFields = "") {
+export async function getByID(id, additionalFields = "") {
     if (!id) throw new ServiceError("user-by-id");
 
     const fields = additionalFields ? additionalFields.split(" ") : [];
@@ -180,7 +180,7 @@ async function getByID(id, additionalFields = "") {
  * @param {{encrypted: boolean}=} options
  * @returns {Promise<User>?|Error>}
  */
-async function getByEmail(email, additionalFields = "", options = { encrypted: false }) {
+export async function getByEmail(email, additionalFields = "", options = { encrypted: false }) {
     if (!email) throw new ServiceError("user-by-email");
     if (!options.encrypted) email = encrypt(email);
 
@@ -542,75 +542,5 @@ export async function registerTutor(userID, attributes) {
         }
     } catch (err) {
         throw new ServiceError("user-update");
-    }
-}
-
-/**
- * Tutee requests for a tutor for selected subject(s)
- * @param {string} userID User ID
- * @param {string} tutorID Tutor ID
- * @param {number[]} subjects List of subject IDs
- * @returns {{success: true, message: string}}
- */
-export async function requestTutor(userID, tutorID, subjects = []) {
-    if (!userID || !tutorID || !subjects.length)
-        throw new ServiceError("user-request-tutor-missing");
-
-    const user = await getByID(userID);
-    const tutor = await getByID(tutorID, "is_tutor subjects");
-
-    if (!user || !tutor) throw new ServiceError("user-tutor-not-found");
-    if (!tutor.is_tutor) throw new ServiceError("user-not-tutor");
-    if (userID === tutorID) throw new ServiceError("user-tutor-same");
-
-    // check if subjects being requested is offered by tutor
-    const notOffered = subjects.some(subject => !tutor.subjects.includes(subject));
-    const requestID = `${userID}:${tutorID}`
-    if (notOffered) throw new ServiceError("user-tutor-subject-unoffered");
-
-    // check already similar request
-    const { rows: requests } = await query(
-        "SELECT * FROM tutee_tutor_relationship WHERE id = $1",
-        [requestID]
-    );
-
-    if (requests.length) {
-        const { subjects: tutoredSubjects } = requests[0];
-        const requestedSubjects = subjects.length === tutoredSubjects.length &&
-            subjects.every(subject => tutoredSubjects.includes(subject));
-
-        if (!requestedSubjects) {
-            // update subjects
-            // TODO: email tutor of the change of subjects
-
-            await query(
-                `UPDATE tutee_tutor_relationship SET subjects = $1 WHERE id = $2`,
-                [subjects, requestID]
-            )
-
-            return {
-                success: true,
-                message: "Updated tuition subjects. Your tutor is notified of the changes"
-            }
-        } else {
-            // throw duplicate error
-            throw new ServiceError("user-request-tutor-unique")
-        }
-    };
-
-    // TODO: email request to tutor: tutor to accept/decline
-    // change status of relationship if accepted, delete row if decline
-
-    const queryText = `
-        INSERT INTO tutee_tutor_relationship(tutee_id, tutor_id, subjects)
-        VALUES($1, $2, $3) RETURNING *
-    `
-
-    const queryValues = [userID, tutorID, subjects];
-    await query(queryText, queryValues)
-
-    return {
-        success: true,
-        message: "Request for tuition sent to tutor"
     }
 }
